requirements.txt
asgiref==3.7.2
blinker==1.7.0
click==8.1.7
colorama==0.4.6
colormath==3.0.0
contourpy==1.2.0
cycler==0.12.1
Django==5.0.1
Flask==3.0.0
Flask-Reuploaded==1.4.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.47.0
greenlet==3.0.3
image==1.5.33
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
networkx==3.2.1
numpy==1.26.3
packaging==23.2
pillow==10.2.0
Pylette==0.4
pyparsing==3.1.1
python-dateutil==2.8.2
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
threadpoolctl==3.2.0
typing_extensions==4.9.0
tzdata==2023.4
Werkzeug==3.0.1
WTForms==3.1.2
